Move UserDetailModal to TitleBar.tsx
